Remove last bits of rx from the bot.ts file
